tests/integration/AutoMLOps/services/submission_service/requirements.txt
kfp<2.0.0
google-cloud-aiplatform
google-cloud-logging
Flask
functions-framework==3.*